+++ eras/byron/crypto/test/Test/Cardano/Crypto/Signing/Redeem.hs
@@ -46,9 +46,9 @@ prop_redeemSign = property $ do
   (vk, sk) <- forAll genRedeemKeypair
   a <- forAll genData
 
-  assert $
-    verifyRedeemSig Dummy.protocolMagicId SignForTestingOnly vk a $
-      redeemSign Dummy.protocolMagicId SignForTestingOnly sk a
+  assert
+    $ verifyRedeemSig Dummy.protocolMagicId SignForTestingOnly vk a
+    $ redeemSign Dummy.protocolMagicId SignForTestingOnly sk a
 
 -- | Signing fails when the wrong 'RedeemVerificationKey' is used
 prop_redeemSignDifferentKey :: Property