+++ eras/byron/ledger/impl/testlib/Test/Cardano/Chain/Config.hs
@@ -11,12 +11,7 @@ import Cardano.Crypto.Hashing (Hash, decodeHash)
 import Cardano.Crypto.ProtocolMagic (RequiresNetworkMagic (..))
 import Cardano.Crypto.Raw (Raw)
 import Cardano.Prelude
-
-#ifdef CARDANO_LEDGER_BYRON_TEST
-import Paths_cardano_ledger_byron_test (getDataFileName)
-#else
 import Paths_cardano_ledger_byron (getDataFileName)
-#endif
 
 -- | Read the test mainnet configuration file from the @test@ directory.
 --
@@ -26,7 +21,7 @@ import Paths_cardano_ledger_byron (getDataFileName)
 -- We use `RequiresNoMagic`, as it indicates mainnet
 readMainetCfg :: MonadIO m => m Genesis.Config
 readMainetCfg = do
-  mainnetGenesisJson <- liftIO $ getDataFileName "mainnet-genesis.json"
+  mainnetGenesisJson <- liftIO $ getDataFileName "test/mainnet-genesis.json"
   let genHash =
         either
           (panic . show . Genesis.GenesisHashDecodeError)