+++ eras/byron/ledger/impl/testlib/Test/Cardano/Chain/MempoolPayload/CBOR.hs
@@ -23,7 +23,6 @@ import Cardano.Ledger.Binary (
  )
 import Cardano.Prelude
 import qualified Data.ByteString.Lazy as LBS
-import GetDataFileName ((<:<))
 import Hedgehog (Property, tripping)
 import Test.Cardano.Chain.MempoolPayload.Example (
   exampleMempoolPayload,
@@ -93,49 +92,49 @@ goldenMempoolPayload :: Property
 goldenMempoolPayload =
   goldenTestCBOR
     exampleMempoolPayload
-    <:< "golden/cbor/mempoolpayload/MempoolPayload"
+    "golden/cbor/mempoolpayload/MempoolPayload"
 
 goldenMempoolPayloadFilledIn :: Property
 goldenMempoolPayloadFilledIn =
   filledInGoldenTestCBOR
     exampleMempoolPayload
-    <:< "golden/cbor/mempoolpayload/MempoolPayload"
+    "golden/cbor/mempoolpayload/MempoolPayload"
 
 goldenMempoolPayload1 :: Property
 goldenMempoolPayload1 =
   goldenTestCBOR
     exampleMempoolPayload1
-    <:< "golden/cbor/mempoolpayload/MempoolPayload1"
+    "golden/cbor/mempoolpayload/MempoolPayload1"
 
 goldenMempoolPayload1FilledIn :: Property
 goldenMempoolPayload1FilledIn =
   filledInGoldenTestCBOR
     exampleMempoolPayload1
-    <:< "golden/cbor/mempoolpayload/MempoolPayload1"
+    "golden/cbor/mempoolpayload/MempoolPayload1"
 
 goldenMempoolPayload2 :: Property
 goldenMempoolPayload2 =
   goldenTestCBOR
     exampleMempoolPayload2
-    <:< "golden/cbor/mempoolpayload/MempoolPayload2"
+    "golden/cbor/mempoolpayload/MempoolPayload2"
 
 goldenMempoolPayload2FilledIn :: Property
 goldenMempoolPayload2FilledIn =
   filledInGoldenTestCBOR
     exampleMempoolPayload2
-    <:< "golden/cbor/mempoolpayload/MempoolPayload2"
+    "golden/cbor/mempoolpayload/MempoolPayload2"
 
 goldenMempoolPayload3 :: Property
 goldenMempoolPayload3 =
   goldenTestCBOR
     exampleMempoolPayload3
-    <:< "golden/cbor/mempoolpayload/MempoolPayload3"
+    "golden/cbor/mempoolpayload/MempoolPayload3"
 
 goldenMempoolPayload3FilledIn :: Property
 goldenMempoolPayload3FilledIn =
   filledInGoldenTestCBOR
     exampleMempoolPayload3
-    <:< "golden/cbor/mempoolpayload/MempoolPayload3"
+    "golden/cbor/mempoolpayload/MempoolPayload3"
 
 ts_roundTripMempoolPayload :: TSProperty
 ts_roundTripMempoolPayload =