+++ eras/byron/ledger/impl/CHANGELOG.md
@@ -1,8 +1,12 @@
 # Revision history for `cardano-ledger-byron`
 
-## 1.1.0.1
+## 1.2.0.0
 
-*
+### `testlib`
+
+* Added `testlib` sublibrary as replacement of `cardano-ledger-byron-test` package
+* Depend on libraries that use crypton instead of cryptonite (from former `cardano-ledger-byron-test` v1.5.2.1)
+* Move `mainnet-genesis.json` to this test package to make `ts_prop_elaboratedCertsValid` pass when run with `cabal`. #4586 (from former `cardano-ledger-byron-test` v1.5.2.0)
 
 ## 1.1.0.0